Move DiceView.svelte into tabletop v2 components
It's the only file left outside _components/v2/ after the v1 removal;
relocating it makes the directory read as v2-only.
